pipeline_ecommerce/requirements.txt
requests
pandas
kafka-python
psycopg2-binary
pyspark
python-dotenv
